archer skills in correct place
