+++ src/map/skills/ac/skill_ac_double.cpp
@@ -2,3 +2,11 @@
 
 SkillAC_DOUBLE::SkillAC_DOUBLE() : WeaponSkillImpl(AC_DOUBLE) {
 }
+
+void SkillAC_DOUBLE::calculateSkillRatio(const Damage* wd, const block_list* src, const block_list* target, uint16 skill_lv, int32& base_skillratio) const {
+	base_skillratio += 10 * (skill_lv - 1);
+}
+
+void SkillAC_DOUBLE::castendDamageId(block_list* src, block_list* target, uint16 skill_lv, t_tick tick, int32 flag) const {
+	skill_attack(BF_WEAPON, src, src, target, getSkillId(), skill_lv, tick, flag);
+}
