+++ src/map/skills/ac/skill_ac_owl.hpp
@@ -1,10 +1,9 @@
 #pragma once
 
 #include "../skill_impl.hpp"
-
 #include "../../battle.hpp"
 
 class SkillAC_OWL : public SkillImpl {
 public:
-	SkillAC_OWL();
+    SkillAC_OWL();
 };
