+++ src/map/skills/ac/skill_ac_vulture.hpp
@@ -1,10 +1,9 @@
 #pragma once
 
 #include "../skill_impl.hpp"
-
 #include "../../battle.hpp"
 
 class SkillAC_VULTURE : public SkillImpl {
 public:
-	SkillAC_VULTURE();
+    SkillAC_VULTURE();
 };
