--- src/map/skills/skill_ac_concentration.hpp
@@ -1,11 +1,0 @@
-#pragma once
-
-#include "../weapon_skill_impl.hpp"
-#include "../../battle.hpp"
-
-class SkillAC_CONCENTRATION : public WeaponSkillImpl {
-public:
-    SkillAC_CONCENTRATION();
-    
-    void castendNoDamageId(block_list* src, block_list* target, uint16 skill_lv, t_tick tick, int32 flag) const override;
-};
