--- src/map/skills/skill_ac_owl.cpp
@@ -1,4 +1,0 @@
-#include "skill_ac_owl.hpp"
-
-SkillAC_OWL::SkillAC_OWL() : SkillImpl(AC_OWL) {
-}
