--- src/map/skills/skill_ac_owl.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "../skill_impl.hpp"
-#include "../../battle.hpp"
-
-class SkillAC_OWL : public SkillImpl {
-public:
-    SkillAC_OWL();
-};
