--- src/map/skills/skill_ac_vulture.cpp
@@ -1,4 +1,0 @@
-#include "skill_ac_vulture.hpp"
-
-SkillAC_VULTURE::SkillAC_VULTURE() : SkillImpl(AC_VULTURE) {
-}
