--- src/map/skills/skill_ac_vulture.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#include "../skill_impl.hpp"
-#include "../../battle.hpp"
-
-class SkillAC_VULTURE : public SkillImpl {
-public:
-    SkillAC_VULTURE();
-};
